Add a service layer and implement business logic.
A service layer package has been created for storing the
implementation of business logic.
The RoomOccupancyService interface has been added with
the method of 'optimize' that the RoomOccupancyServiceImpl
class implementation. The method optimize method consist of the
private methods. The 'distributePremiumGuests' private method
responsible for adding all premium guests to the available
premium rooms. The 'distributeRestOfTheGuests' private
method stands for distributing economy guests based on
available economy rooms and premium rooms.
If all economy guests can fit in with the available economy
rooms, then it puts all of them into economy rooms.
Otherwise, it puts the highest payable quest(s) into the
premium rooms, and then to the economy rooms.